pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "grantsage"
version = "0.1.0"
description = "GrantSage - Copiloto para Edital e Proposta"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.28.0",
    "python-dotenv>=1.0.0",
    "pymupdf>=1.24.0",
    "mammoth>=1.8.0",
    "markdownify>=0.13.0",
    "beautifulsoup4>=4.12.0",
    "pypandoc>=1.13.0",
    "litellm>=1.40.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py310"]
extend-exclude = "/(\\.venv|\\.git|build|dist)/"